Compute the winning threshold using Bayesian Optimization

This is a simple wrapper for rBayesianOptimization::BayesianOptimization that seeks to discover a threshold that produces a small-world, scale-free network over a given range. Generally 10 or more iterations (including init points) is needed to find an optimal value. For smaller networks ( under 10000 genes), it should be no problem to run 20 or more iterations within a few minutes.

bayes_thresh(
  mm,
  bounds,
  init_grid_dt = NULL,
  init_points = 0,
  n_iter,
  acq = "ucb",
  kappa = 2.576,
  eps = 0,
  kernel = list(type = "exponential", power = 2),
  verbose = FALSE,
  ...
)

Arguments

mm
A metric matrix or an object of class 'dist'

bounds
A numeric vector of lower and upper bounds. All the sample points in init_grid_dt should be in the range of bounds.

init_grid_dt
A numeric vector of user specified points to sample the target function.

init_points
Number of randomly chosen points to sample the target function before Bayesian Optimization fitting the Gaussian Process.

n_iter
Total number of times the Bayesian Optimization is to repeated.

acq
Acquisition function type to be used. Can be "ucb", "ei" or "poi".

ucb GP Upper Confidence Bound

ei Expected Improvement

poi Probability of Improvement

kappa
tunable parameter kappa of GP Upper Confidence Bound, to balance exploitation against exploration, increasing kappa will make the optimized hyperparameters pursuing exploration.

eps
tunable parameter epsilon of Expected Improvement and Probability of Improvement, to balance exploitation against exploration, increasing epsilon will make the optimized hyperparameters are more spread out across the whole range.

kernel
Kernel (aka correlation function) for the underlying Gaussian Process. This parameter should be a list that specifies the type of correlation function along with the smoothness parameter. Popular choices are square exponential (default) or matern 5/2

verbose
Whether or not to print progress.

...
Other arguments passed on to GP_fit.

Value

a list of Bayesian Optimization result is returned:

Best_Par a named vector of the best hyperparameter set found

Best_Value the value of metrics achieved by the best hyperparameter set

History a data.table of the bayesian optimization history

Pred a data.table with validation/cross-validation prediction for each round of bayesian optimization history